+++ src/main/java/edu/miu/lab2/controller/CourseController.java
@@ -1,7 +1,7 @@
-package edu.miu.restfuldemo.controller;
+package edu.miu.lab2.controller;
 
-import edu.miu.restfuldemo.dto.CourseDto;
-import edu.miu.restfuldemo.service.CourseService;
+import edu.miu.lab2.dto.CourseDto;
+import edu.miu.lab2.service.CourseService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
