+++ src/main/java/edu/miu/lab2/repo/StudentRepo.java
@@ -1,12 +1,11 @@
-package edu.miu.restfuldemo.repo;
+package edu.miu.lab2.repo;
 
-import edu.miu.restfuldemo.entity.Student;
+import edu.miu.lab2.entity.Student;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @Repository
