+++ src/main/java/edu/miu/lab2/service/CourseService.java
@@ -1,7 +1,6 @@
-package edu.miu.restfuldemo.service;
+package edu.miu.lab2.service;
 
-import edu.miu.restfuldemo.dto.CourseDto;
-import edu.miu.restfuldemo.entity.Course;
+import edu.miu.lab2.dto.CourseDto;
 
 import java.util.List;
