+++ src/test/java/edu/miu/lab2/Lab2ApplicationTests.java
@@ -1,4 +1,4 @@
-package edu.miu.restfuldemo;
+package edu.miu.lab2;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
